FModSound pass (wave 3, MCP-free): FMOD vtable-dispatch static wrappers replace 5 spurious PLT imports (1095->1090, matches original import set); promptMusicCue/setMusicParamValue/setVolume linked-exact; stopAllSoundFXEvents 65->85.7; downPitch int8_t (strb). byte 1122->1125, linked 2200->2201. Trade: updateAll 34.6->31.0 requeued. RATCHET PASS.

+++ src/engine/audio/FModSound.h
@@ -13,7 +13,7 @@ class FModSound {
 public:
     int currentMusicEvent;
     int fadeTargetMusicEvent;
-    int downPitch;
+    int8_t downPitch;
     char *appRootDir;
     uint8_t lowMemory;
     uint8_t categoryEnabled[4];
